requirements.txt
Flask==1.1.2
Flask-RESTful
Flask-SQLAlchemy
flask-marshmallow 
marshmallow-sqlalchemy 
marshmallow 
asyncio-nats-streaming
pytest==6.0.1
pytest-cov==2.10.0
coverage==5.2.1